bhyve: Move the RTC driver to amd64/
No functional change intended.

Reviewed by:	corvink, jhb
MFC after:	1 week
Sponsored by:	Innovate UK
Differential Revision:	https://reviews.freebsd.org/D40734

(cherry picked from commit a7f6c2ff732db1d0032066109ee0dd5af1a39b81)
