+++ usr.sbin/bhyve/amd64/rtc.c
