+++ usr.sbin/bhyve/amd64/rtc.h
